#6 rename: 통계 데이터 조회 Path 변경
